Move branch extensions to OperationExtensions.cs
